add: runner_socket. fix: compile error.

+++ src/balancer/io/prio_balancer.cpp
@@ -4,8 +4,8 @@ namespace lon::io {
 
 PrioBlancer::PrioBlancer(std::vector<uint8_t> prio_threads_count) {
     prio_workers.resize(prio_threads_count.size());
-    for (int i = 0; i < prio_threads_count.size(); ++i) {
-        for (int count = 0; count < prio_threads_count[i]; ++count)
+    for (size_t i = 0; i < prio_threads_count.size(); ++i) {
+        for (size_t count = 0; count < prio_threads_count[i]; ++count)
             threads_.emplace_back([&]() {
                 auto manager = IOManager::getThreadLocal();
                 prio_workers[i].push_back(manager);
@@ -24,8 +24,8 @@ void PrioBlancer::schedule(coroutine::Executor::Ptr executor, std::any arg) {
     }
 
     //超过注册的最大优先级, 降级执行.
-    if (prio > prio_workers.size() - 1)
-        prio = prio_workers.size() - 1;
+    if (prio > static_cast<int>(prio_workers.size() - 1))
+        prio = static_cast<int>(prio_workers.size() - 1);
     //当前优先级没有线程, 降级执行.
     while (prio_workers[prio].size() == 0 && prio >= 0) {
         --prio;
